[move-compiler] Add check_constant_visibility and fold constant function generation into typing translation
