+++ external-crates/move/crates/move-compiler/src/typing/mod.rs
@@ -7,7 +7,6 @@ pub mod core;
 mod dependency_ordering;
 mod deprecation_warnings;
 mod expand;
-mod generate_constant_functions;
 mod infinite_instantiations;
 mod macro_expand;
 mod match_analysis;
